Update Java version and dependencies across project
Updated the Java version from 19 to 21 in Gradle, GitHub workflows, and SDKMAN config files. Multiple dependencies were updated in the build.gradle file and a method invocation was changed in MindergasnlService.java to match the new library version. The Gradle version was also updated. These changes were necessary to maintain compatibility with the latest Java features and remain up to date with the newest versions of dependencies.