Enhance H2 DB settings and documentation
Updated configurations in application-dev-h2db.properties for H2 database to run in MySQL mode and enable always defragmentation. This aids in better performance and maintenance of the database.

Additionally, enabled H2 console settings for direct SQL access and manipulation. This adds more robust options for database management.

Updated README documentation with steps to compact the H2 database. This action is often needed during the lifecycle of the database to release unused space and defragment the file, thus improving performance.